Made Program Async
Executed calls to `test_site` in the background. This works fine but the
program instantly returns if there is no wait statement. Therefore,
recorded each test's pid into a list with a counter. This list is then
iterated and each process is waited on. The program will not return
until all sites have been reported as successful or failures.